refactor(playground): extract shared AgentSelectableCard
Models and Tools rendered nearly identical selectable cards (agent-color
container/check styling, focus/hover states, optional left visual,
title/subtitle, check mark). Extract that shared markup into
AgentSelectableCard at the agent-edit level and use it directly at the
map call sites — no purposeless ModelListEntry/ToolItem wrapper layer.
