+++ packages/playground/src/domains/agent-builder/components/agent-edit/agent-profile/tools.tsx
@@ -1,12 +1,11 @@
-import { Txt, cn } from '@mastra/playground-ui';
-import { Check } from 'lucide-react';
-import type { CSSProperties, ReactNode } from 'react';
+import { Txt } from '@mastra/playground-ui';
+import type { ReactNode } from 'react';
 import { useState } from 'react';
 import { useFormContext } from 'react-hook-form';
-import { useAgentColor } from '../../../contexts/agent-color-context';
 import type { AgentBuilderEditFormValues } from '../../../schemas';
 import type { AgentTool } from '../../../types/agent-tool';
 import { AgentSearchbar } from '../agent-searchbar';
+import { AgentSelectableCard } from '../agent-selectable-card';
 
 interface ToolsProps {
   editable?: boolean;
@@ -52,90 +51,21 @@ export const Tools = ({ editable = true, availableAgentTools = [] }: ToolsProps)
       ) : (
         <div className="grid min-h-0 grid-cols-1 gap-2 lg:gap-6 overflow-y-auto sm:grid-cols-2 lg:grid-cols-3">
           {visibleTools.map(item => (
-            <ToolItem key={`${item.type}__${item.id}`} item={item} editable={editable} onToggle={toggle} />
+            <AgentSelectableCard
+              key={`${item.type}__${item.id}`}
+              title={item.name}
+              subtitle={item.description || 'No description provided'}
+              isSelected={item.isChecked}
+              disabled={!editable}
+              onClick={() => toggle(item, !item.isChecked)}
+              ariaLabel={item.name}
+              testId={`tool-card-${item.type}-${item.id}`}
+              checkTestId={`tool-card-check-${item.type}-${item.id}`}
+            />
           ))}
         </div>
       )}
     </div>
-  );
-};
-
-interface ToolItemProps {
-  item: AgentTool;
-  editable: boolean;
-  onToggle: (item: AgentTool, next: boolean) => void;
-}
-
-const ToolItem = ({ item, editable, onToggle }: ToolItemProps) => {
-  const agentColor = useAgentColor();
-  const hasAgentColor = agentColor !== null;
-  const useAgentColors = item.isChecked && hasAgentColor;
-
-  const containerStyle: CSSProperties | undefined = hasAgentColor
-    ? {
-        ['--agent-color-bg' as string]: agentColor.background,
-        ...(item.isChecked ? { borderColor: agentColor.background } : null),
-      }
-    : undefined;
-
-  const checkStyle: CSSProperties | undefined = useAgentColors
-    ? {
-        borderColor: agentColor.background,
-        backgroundColor: agentColor.background,
-        color: agentColor.foreground,
-      }
-    : undefined;
-
-  return (
-    <button
-      key={`${item.type}__${item.id}`}
-      type="button"
-      onClick={() => onToggle(item, !item.isChecked)}
-      disabled={!editable}
-      aria-pressed={item.isChecked}
-      aria-label={item.name}
-      data-testid={`tool-card-${item.type}-${item.id}`}
-      style={containerStyle}
-      className={cn(
-        'flex items-center gap-3 rounded-lg border bg-surface3 p-4 text-left transition-colors cursor-pointer active:opacity-90',
-        hasAgentColor
-          ? 'focus-visible:!border-[var(--agent-color-bg)] focus-visible:outline-none'
-          : 'hover:bg-surface4 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-accent1',
-        hasAgentColor && 'hover:bg-surface4',
-        item.isChecked
-          ? useAgentColors
-            ? 'bg-surface4'
-            : 'border-accent1 bg-surface4 ring-1 ring-accent1'
-          : 'border-border1',
-        !editable && 'cursor-not-allowed opacity-60',
-      )}
-    >
-      <div className="flex min-w-0 flex-1 flex-col">
-        <Txt variant="ui-md" className="truncate font-medium text-neutral6">
-          {item.name}
-        </Txt>
-
-        <Txt variant="ui-sm" className="truncate text-neutral3">
-          {item.description || 'No description provided'}
-        </Txt>
-      </div>
-
-      <span
-        aria-hidden="true"
-        data-testid={`tool-card-check-${item.type}-${item.id}`}
-        style={checkStyle}
-        className={cn(
-          'flex h-4 w-4 shrink-0 items-center justify-center rounded border transition-colors',
-          item.isChecked
-            ? useAgentColors
-              ? ''
-              : 'border-accent1 bg-accent1 text-surface1'
-            : 'border-border1 bg-transparent',
-        )}
-      >
-        {item.isChecked && <Check className="h-3 w-3" />}
-      </span>
-    </button>
   );
 };
